Выполнение задания 7 - не завершил
