+++ attack.js
@@ -1,0 +1,5 @@
+function attack() {
+    console.log(this.name + ' ' + ' Fight...');
+};
+
+export default attack;
