+++ createPlayer.js
@@ -1,0 +1,24 @@
+import createElement from './createElement.js';
+
+function createPlayer (playerInfo) {
+    const $player = createElement('div', 'player' + playerInfo.player);
+    const $progressbar = createElement('div', 'progressbar');
+    const $character = createElement('div', 'character');
+    const $life = createElement('div', 'life');
+    const $name = createElement('div', 'name');
+    const $img = createElement('img');
+
+    $life.style.width = playerInfo.hp + '%';
+    $name.innerText = playerInfo.name;
+    $img.src = playerInfo.img;
+
+    $character.appendChild($img);
+    $progressbar.appendChild($life);
+    $progressbar.appendChild($name);
+    $player.appendChild($progressbar);
+    $player.appendChild($character);
+
+    return $player;
+};
+
+export default createPlayer;
