+++ enemyAttack.js
@@ -1,0 +1,10 @@
+function enemyAttack () {
+    const hit = ATTACK[getRandom(3) - 1];
+    const defence = ATTACK[getRandom(3) - 1];
+
+    return {
+        value: getRandom(HIT[hit]),
+        hit,
+        defence,
+    }
+};
