+++ fightButtonOff.js
@@ -1,0 +1,7 @@
+function fightButtonOff () {
+    if (player1.hp === 0 || player2.hp === 0) {
+        $randomButton.disabled = true;
+        $randomButton.style.backgroundColor = 'grey';
+        createReloadButton();
+    }
+};
