+++ main.js
@@ -1,7 +1,15 @@
-const $arenas = document.querySelector('.arenas');
-const $randomButton = document.querySelector('.button');
-const $formFight = document.querySelector('.control');
-const $chat = document.querySelector('.chat');
+import { player1 } from './player1.js';
+import { player2 } from './player2.js';
+
+import createPlayer from './createPlayer.js';
+import createReloadButton from './createReloadButton.js';
+import playerWins from './playerWins.js';
+import renderMessage from './renderMessage.js';
+
+import { $arenas } from './const.js';
+import { $randomButton } from './const.js';
+import { $formFight } from './const.js';
+import { $chat } from './const.js';
 
 const HIT = {
     head: 30,
@@ -11,221 +19,10 @@ const HIT = {
 
 const ATTACK = ['head', 'body', 'foot'];
 
-const logs = {
-    start: 'Часы показывали [time], когда [player1] и [player2] бросили вызов друг другу.',
-    end: [
-        'Результат удара [playerWins]: [playerLose] - труп',
-        '[playerLose] погиб от удара бойца [playerWins]',
-        'Результат боя: [playerLose] - жертва, [playerWins] - убийца',
-    ],
-    hit: [
-        '[playerDefence] пытался сконцентрироваться, но [playerKick] разбежавшись раздробил копчиком левое ухо врага.',
-        '[playerDefence] расстроился, как вдруг, неожиданно [playerKick] случайно раздробил грудью грудину противника.',
-        '[playerDefence] зажмурился, а в это время [playerKick], прослезившись, раздробил кулаком пах оппонента.',
-        '[playerDefence] чесал <вырезано цензурой>, и внезапно неустрашимый [playerKick] отчаянно размозжил грудью левый бицепс оппонента.',
-        '[playerDefence] задумался, но внезапно [playerKick] случайно влепил грубый удар копчиком в пояс оппонента.',
-        '[playerDefence] ковырялся в зубах, но [playerKick] проснувшись влепил тяжелый удар пальцем в кадык врага.',
-        '[playerDefence] вспомнил что-то важное, но внезапно [playerKick] зевнув, размозжил открытой ладонью челюсть противника.',
-        '[playerDefence] осмотрелся, и в это время [playerKick] мимоходом раздробил стопой аппендикс соперника.',
-        '[playerDefence] кашлянул, но внезапно [playerKick] показав палец, размозжил пальцем грудь соперника.',
-        '[playerDefence] пытался что-то сказать, а жестокий [playerKick] проснувшись размозжил копчиком левую ногу противника.',
-        '[playerDefence] забылся, как внезапно безумный [playerKick] со скуки, влепил удар коленом в левый бок соперника.',
-        '[playerDefence] поперхнулся, а за это [playerKick] мимоходом раздробил коленом висок врага.',
-        '[playerDefence] расстроился, а в это время наглый [playerKick] пошатнувшись размозжил копчиком губы оппонента.',
-        '[playerDefence] осмотрелся, но внезапно [playerKick] робко размозжил коленом левый глаз противника.',
-        '[playerDefence] осмотрелся, а [playerKick] вломил дробящий удар плечом, пробив блок, куда обычно не бьют оппонента.',
-        '[playerDefence] ковырялся в зубах, как вдруг, неожиданно [playerKick] отчаянно размозжил плечом мышцы пресса оппонента.',
-        '[playerDefence] пришел в себя, и в это время [playerKick] провел разбивающий удар кистью руки, пробив блок, в голень противника.',
-        '[playerDefence] пошатнулся, а в это время [playerKick] хихикая влепил грубый удар открытой ладонью по бедрам врага.',
-    ],
-    defence: [
-        '[playerKick] потерял момент и храбрый [playerDefence] отпрыгнул от удара открытой ладонью в ключицу.',
-        '[playerKick] не контролировал ситуацию, и потому [playerDefence] поставил блок на удар пяткой в правую грудь.',
-        '[playerKick] потерял момент и [playerDefence] поставил блок на удар коленом по селезенке.',
-        '[playerKick] поскользнулся и задумчивый [playerDefence] поставил блок на тычок головой в бровь.',
-        '[playerKick] старался провести удар, но непобедимый [playerDefence] ушел в сторону от удара копчиком прямо в пятку.',
-        '[playerKick] обманулся и жестокий [playerDefence] блокировал удар стопой в солнечное сплетение.',
-        '[playerKick] не думал о бое, потому расстроенный [playerDefence] отпрыгнул от удара кулаком куда обычно не бьют.',
-        '[playerKick] обманулся и жестокий [playerDefence] блокировал удар стопой в солнечное сплетение.'
-    ],
-    draw: 'Ничья - это тоже победа!'
-};
-
-const player1 = {
-    player: 1,
-    name: 'Kitana',
-    hp: 100,
-    img: 'http://reactmarathon-api.herokuapp.com/assets/kitana.gif',
-    weapon: ['Боевые веера'],
-    elHP,
-    changeHP,
-    renderHP,
-    attack,
-};
-
-const player2 = {
-    player: 2,
-    name: 'Sonya',
-    hp: 100,
-    img: 'http://reactmarathon-api.herokuapp.com/assets/sonya.gif',
-    weapon: ['Бамбуковые палки'],
-    elHP,
-    changeHP,
-    renderHP,
-    attack,
-};
-
-function attack() {
-    console.log(this.name + ' ' + ' Fight...');
-};
-
-function createElement(tag, className) {
-    const $tag = document.createElement(tag);
-    if (className) {
-        $tag.classList.add(className);
-    }
-
-    return $tag;
-};
-
-function createPlayer(playerInfo) {
-    const $player = createElement('div', 'player' + playerInfo.player);
-    const $progressbar = createElement('div', 'progressbar');
-    const $character = createElement('div', 'character');
-    const $life = createElement('div', 'life');
-    const $name = createElement('div', 'name');
-    const $img = createElement('img');
-
-    $life.style.width = playerInfo.hp + '%';
-    $name.innerText = playerInfo.name;
-    $img.src = playerInfo.img;
-
-    $character.appendChild($img);
-    $progressbar.appendChild($life);
-    $progressbar.appendChild($name);
-    $player.appendChild($progressbar);
-    $player.appendChild($character);
-
-    return $player;
-};
-
-function elHP () {
-    return document.querySelector('.player' + this.player + ' .life');
-};
-
-function changeHP (hp) {
-    this.hp -= hp;
-
-    if (this.hp <= 0) {
-        this.hp = 0;
-    }
-};
-
-function renderHP () {
-    return this.elHP().style.width = this.hp + '%';
-};
-
-function getRandom (meaning) {
-    return Math.ceil(Math.random() * meaning)
-};
-
-function playerWins (name) {
-    const $winsTitle = createElement('div', 'winsTitle');
-
-    if (name) {
-        $winsTitle.innerText = name + ' wins';
-    } else {
-        $winsTitle.innerText = 'drow';
-    }
-
-    return $winsTitle;
-};
-
-function createReloadButton () {
-    const $reloadWrap = createElement('div', 'reloadWrap');
-    const $resetButton = createElement('button', 'button');
-
-    $resetButton.innerText = 'Reset';
-
-    $resetButton.addEventListener('click', function () {
-        window.location.reload();
-    });
-
-    $reloadWrap.appendChild($resetButton);
-    $arenas.appendChild($reloadWrap);
-};
+const getRandom = (meaning) => Math.ceil(Math.random() * meaning);
 
 $arenas.appendChild(createPlayer(player1));
 $arenas.appendChild(createPlayer(player2));
-
-function enemyAttack () {
-    const hit = ATTACK[getRandom(3) - 1];
-    const defence = ATTACK[getRandom(3) - 1];
-
-    return {
-        value: getRandom(HIT[hit]),
-        hit,
-        defence,
-    }
-};
-
-function renderMessage () {
-    if (player1.hp === 0 && player1.hp < player2.hp) {
-        $arenas.appendChild(playerWins(player2.name));
-        generateLogs('end', player2, player1);
-    } else if (player2.hp === 0 && player2.hp < player1.hp) {
-        $arenas.appendChild(playerWins(player1.name));
-        generateLogs('end', player1, player2);
-    } else if (player1.hp === 0 && player2.hp === 0) {
-        $arenas.appendChild(playerWins());
-        generateLogs('draw');
-    }
-};
-
-function fightButtonOff () {
-    if (player1.hp === 0 || player2.hp === 0) {
-        $randomButton.disabled = true;
-        $randomButton.style.backgroundColor = 'grey';
-        createReloadButton();
-    }
-};
-
-function fight (enemy, player) {
-    if (enemy.hit !== player.defence) {
-        player2.changeHP(enemy.value);
-        player2.renderHP();
-        generateLogs('hit', player1, player2, enemy.value);
-    }  else {
-        generateLogs('defence', player1, player2);
-    }
-
-    if (player.hit !== enemy.defence) {
-        player1.changeHP(player.value);
-        player1.renderHP();
-        generateLogs('hit', player2, player1, player.value);
-    } else {
-        generateLogs('defence', player2, player1);
-    }
-};
-
-function playerAttack () {
-    const attack = {};
-
-    for (let item of $formFight) {
-        if (item.checked && item.name === 'hit') {
-            attack.value = getRandom(HIT[item.value]);
-            attack.hit = item.value;
-        }
-
-        if (item.checked && item.name === 'defence') {
-            attack.defence = item.value;
-        }
-
-        item.checked = false;
-    }
-
-    return attack;
-};
 
 function generateLogs (type, player1, player2, value) {
     const date = new Date();
