+++ player1.js
@@ -1,0 +1,16 @@
+import elHP from './elHP.js';
+import changeHP from './elHP.js';
+import renderHP from './elHP.js';
+import attack from './elHP.js';
+
+export const player1 = {
+    player: 1,
+    name: 'Kitana',
+    hp: 100,
+    img: 'http://reactmarathon-api.herokuapp.com/assets/kitana.gif',
+    weapon: ['Боевые веера'],
+    elHP,
+    changeHP,
+    renderHP,
+    attack,
+};
