+++ renderHP.js
@@ -1,0 +1,5 @@
+function renderHP () {
+    return this.elHP().style.width = this.hp + '%';
+};
+
+export default renderHP;
